requirements.txt
Flask==2.0.1
Flask-JWT-Extended==4.3.1
Flask-CORS==3.0.10
SQLAlchemy==1.4.31
PyMySQL==1.0.2
python-dotenv==0.19.2
Flask-SQLAlchemy==2.5.1
# Add other dependencies as needed